% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{battery}
\alias{battery}
\title{두 회사 건전지 수명}
\format{
A data frame with 45 rows and 2 variables:
\describe{
  \item{\code{회사명}}{integer 회사명, A/B}
  \item{\code{건전지_수명}}{integer 분단위로 측정된 건전지 수명}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
battery
}
\description{
고등학교 실용통계 P68 분단위로 측정된 두회사 건전지 수명
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
