% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{blood_book}
\alias{blood_book}
\title{혈액형과 한학기 동안 읽은 책의 수}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{\code{혈액형}}{integer 혈액형, A/B/AB/O}
  \item{\code{책의수}}{integer 독서 책의 수}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
blood_book
}
\description{
고등학교 실용통계 P59 지홍이네 반 학생 혈액형과 한 학기 동안 읽은 책의 수
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
