% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{cars}
\alias{cars}
\title{국내출시 차량정보}
\format{
A data frame with 30 rows and 2 variables:
\describe{
  \item{\code{중량}}{integer 중량 KG}
  \item{\code{연비}}{integer 연비 리터당 킬로미터}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
cars
}
\description{
고등학교 실용통계 P78 국내 출시 30종 차량의 중량과 연비
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
