% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{grades}
\alias{grades}
\title{영수네반 20명 수학 성적과 과학 성적}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{\code{수학}}{integer 수학성적}
  \item{\code{과학}}{integer 과학성적}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
grades
}
\description{
고등학교 실용통계 P76 영수네반 학생 20명 성적; 수학, 과학
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
