% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{homerun}
\alias{homerun}
\title{두 타자 홈런수}
\format{
A data frame with 34 rows and 2 variables:
\describe{
  \item{\code{타자}}{integer 타자 A,B}
  \item{\code{홈런수}}{integer 홈런수}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
homerun
}
\description{
고등학교 실용통계 P69 A, B 두타자 홈런수
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
