% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{LED}
\alias{LED}
\title{두회사 LED 수명}
\format{
A data frame with 50 rows and 2 variables:
\describe{
  \item{\code{회사명}}{character 회사명 A,B}
  \item{\code{led_수명}}{integer LED 수명}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
LED
}
\description{
A, B 두회사 LED 수명
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
