% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{lengths}
\alias{lengths}
\title{상원이네 반 학생 발길이와 한뼘길이}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{\code{발_길이}}{integer 발길이, mm}
  \item{\code{한_뼘의_길이}}{integer 한뼘길이, mm}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
lengths
}
\description{
고등학교 실용통계 P84 상원이네 반 학생 20명 발 길이와 한 뼘 길이
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
