% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{math_pref}
\alias{math_pref}
\title{수학선호도}
\format{
A data frame with 20 rows and 4 variables:
\describe{
  \item{\code{성별}}{integer 1,2 남성,여성}
  \item{\code{수학_선호도}}{integer 1~5 리커트 척도, 1: 수학 가장 비선호 5: 수학 가장 선호}
  \item{\code{키}}{integer 신장 CM}
  \item{\code{몸무게}}{integer 몸무게 KG}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
math_pref
}
\description{
고등학교 실용통계 P58 수학선호도
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
