% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{measures}
\alias{measures}
\title{현주네반 20명 목둘레와 허리둘레}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{\code{목둘레}}{integer cm 로 측정된 목둘레}
  \item{\code{허리둘레}}{integer cm 로 측정된 허리둘레}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
measures
}
\description{
고등학교 실용통계 P75 현주네반 20명 신체 측정정보; 목둘레, 허리둘레
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
