% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{noodle}
\alias{noodle}
\title{봉지라면 20종 영양성분}
\format{
A data frame with 20 rows and 3 variables:
\describe{
  \item{\code{열량}}{integer 열량, kcal}
  \item{\code{탄수화물}}{integer 탄수화물, gram}
  \item{\code{나트륨}}{integer 나트륨, mg}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
noodle
}
\description{
고등학교 실용통계 P84 봉지라면 20종 영양성분
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
