% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{pullup}
\alias{pullup}
\title{턱걸이 횟수}
\format{
A data frame with 24 rows and 1 variables:
\describe{
  \item{\code{턱걸이_횟수}}{integer 턱걸이 횟수}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
pullup
}
\description{
고등학교 실용통계 P73 턱걸이 횟수
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
