% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{reading}
\alias{reading}
\title{미희네반 학생 독서와 게임시간}
\format{
A data frame with 21 rows and 2 variables:
\describe{
  \item{\code{독서_시간}}{integer 독서시간}
  \item{\code{게임_시간}}{integer 게임시간}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
reading
}
\description{
고등학교 실용통계 P77 미희네반 학생들 하루평균 독서 및 게임시간
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
