% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{scores}
\alias{scores}
\title{농구점수와 축구점수}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{\code{농구_점수}}{integer 농구점수}
  \item{\code{축구_점수}}{integer 축구점수}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
scores
}
\description{
고등학교 실용통계 P74 농구점수와 축구점수
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
