% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_redshift_connection.R
\name{set_redshift_credentials}
\alias{set_redshift_credentials}
\title{Set Redshift Credentials
Create a file that contains information to connect to redshift}
\usage{
set_redshift_credentials(username, password, host, port)
}
\description{
Set Redshift Credentials
Create a file that contains information to connect to redshift
}

