% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{shoes}
\alias{shoes}
\title{한국 고등학생 신발구매 실태}
\format{
A data frame with 60 rows and 6 variables:
\describe{
  \item{\code{학년}}{integer 학년, 1/2/3}
  \item{\code{성별}}{integer 성별, 1:남자/2:여자}
  \item{\code{신발_사이즈}}{integer 신발 사이즈}
  \item{\code{구매_장소}}{integer 구매장소, 1:개별 매장/2:전문 매장/3:백화점/4:할인 마트/5:온라인 매장}
  \item{\code{구매_이유}}{integer 구매이유, 1:가격이 적당해서/2:발이 편해서/3:디자인이 좋아서/4:지인이 권유해서}
  \item{\code{x12시_발길이}}{integer 12시 발길이}
  \item{\code{x18시_발길이}}{integer 18시 발길이}
  \item{\code{적합성}}{integer 0:부적합/1:적합}
}
}
\source{
통계청 통계교육원, "고등학교 실용통계"
}
\usage{
shoes
}
\description{
고등학교 실용통계 P156 한국고등학생 신발구매실태 조사
}
\details{
통그라미 내장 데이터
}
\keyword{datasets}
